requirements.txt
Django==5.0.6
djangorestframework==3.15.2
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
psycopg==3.2.1
aiohttp==3.9.1
tenacity==8.2.3
